prelogin router, friend router...
